drf_noegest/requirements.txt
Django==4.2.3
djangorestframework==3.12.4
requests==2.31.0
djangorestframework_simplejwt==5.2.2
mysqlclient==2.2.6